Change overlay opacity to double (range 0 to 1).

+++ src/strip/BufferedStrip.h
@@ -9,10 +9,10 @@ class BufferedStrip : public StatefulStrip {
     private:
         CRGBSet *leds;
         Strip *strip;
-        uint8_t opacity;
+        double opacity;
 
     public:
-        BufferedStrip(Strip *strip, uint8_t opacity);
+        BufferedStrip(Strip *strip, double opacity);
         void flush() override;
 };
